decodyImportScript/requirements.txt
PyMySQL==1.1.1
pyyaml==6.0.2
python-dotenv==1.0.1
